requirements.txt
django==1.7.8
django-crispy-forms==1.4.0